+++ packages/atlas/src/components/_overlays/MemberDropdown/MemberDropdown.styles.ts
@@ -1,15 +1,8 @@
-import { css } from '@emotion/react'
+import isPropValid from '@emotion/is-prop-valid'
 import styled from '@emotion/styled'
 import { animated } from 'react-spring'
 
-import { Avatar } from '@/components/Avatar'
-import { Text } from '@/components/Text'
-import { IconWrapper } from '@/components/_icons/IconWrapper'
 import { cVar, media, sizes, zIndex } from '@/styles'
-
-const paddingStyles = css`
-  padding: ${sizes(6)} ${sizes(4)};
-`
 
 export const Container = styled.div`
   position: fixed;
@@ -24,45 +17,32 @@ export const Container = styled.div`
   }
 `
 
-type FixedSizeContainerProps = {
-  height?: number | string
-  width?: number | string
-}
-export const FixedSizeContainer = styled.div<FixedSizeContainerProps>`
-  height: ${({ height }) => (height ? (typeof height === 'number' ? `${height}px` : height) : 'unset')};
-  width: ${({ width }) => (width ? (typeof width === 'number' ? `${width}px` : width) : 'unset')};
-`
-
-export const LIST_TRANSITION = 'list-transition'
-
 export const SectionContainer = styled.div`
   border-top: 1px solid ${cVar('colorBorderMutedAlpha')};
   padding: ${sizes(2)} 0;
 `
 
-export const AnimatedSectionContainer = styled(animated(SectionContainer))`
-  border-top: unset;
-  width: 100%;
-`
-
-export const SlideAnimationContainer = styled(animated.div)`
+export const SlideAnimationContainer = styled(animated.div, { shouldForwardProp: isPropValid })<{
+  disableVerticalScroll: boolean
+}>`
   position: absolute;
   height: 100%;
   width: 280px;
   will-change: transform, opacity;
   overflow-x: hidden;
+  overflow-y: ${({ disableVerticalScroll }) => (disableVerticalScroll ? 'hidden' : 'auto')};
 `
 
 export const InnerContainer = styled.div<{
   isActive: boolean
-  containerHeight: number
+  containerHeight?: number
 }>`
   width: 280px;
   position: relative;
   max-height: calc(100vh - ${sizes(4)} - var(--size-topbar-height));
-  height: ${({ containerHeight }) => containerHeight}px;
+  height: ${({ containerHeight = 0 }) => containerHeight}px;
   transform: translateY(
-    ${({ isActive, containerHeight }) =>
+    ${({ isActive, containerHeight = 0 }) =>
       isActive ? 'var(--size-topbar-height)' : `calc(-${containerHeight}px + var(--size-topbar-height)) `}
   );
   transition: transform ${cVar('animationTransitionMedium')}, height ${cVar('animationTransitionMedium')};
@@ -70,134 +50,5 @@ export const InnerContainer = styled.div<{
   box-shadow: ${cVar('effectElevation24Layer2')}, ${cVar('effectElevation24Layer1')};
   border-radius: 0 0 ${cVar('radiusMedium')} ${cVar('radiusMedium')};
   background-color: ${cVar('colorBackgroundStrong')};
-  overflow-y: auto;
   overflow-x: hidden;
 `
-
-export const AvatarsGroupContainer = styled.div`
-  display: grid;
-  justify-content: start;
-  grid-template-columns: auto auto;
-  gap: ${sizes(4)};
-  margin-bottom: ${sizes(6)};
-`
-
-export const AvatarWrapper = styled.div`
-  position: relative;
-`
-
-export const StyledIconWrapper = styled(IconWrapper)`
-  position: absolute;
-  right: -8px;
-  bottom: -8px;
-`
-
-export const StyledAvatar = styled(Avatar)<{ isDisabled: boolean }>`
-  width: 40px;
-  height: 40px;
-  clip-path: path(
-    'm 40 0 h -40 v 40 h 20.5041 c -0.3291 -1.2785 -0.5041 -2.6188 -0.5041 -4 c 0 -8.8366 7.1634 -16 16 -16 c 1.3812 0 2.7215 0.175 4 0.5041 v -20.5041 z'
-  );
-  opacity: ${({ isDisabled }) => (isDisabled ? 0.25 : 1)};
-  transition: opacity ${cVar('animationTransitionMedium')};
-`
-
-export const AddAvatar = styled.div`
-  width: 100%;
-  height: 100%;
-  background-color: ${cVar('colorBackgroundAlpha')};
-  display: flex;
-  align-items: center;
-  justify-content: center;
-`
-
-export const BlurredBG = styled.div<{ url?: string | null }>`
-  position: relative;
-  width: 280px;
-  height: 100%;
-
-  &::before {
-    transition: opacity ${cVar('animationTransitionMedium')};
-    position: absolute;
-    width: inherit;
-    height: inherit;
-    background-image: url(${({ url }) => url});
-    background-size: cover;
-    background-repeat: no-repeat;
-    background-position: center;
-    content: '';
-  }
-`
-
-export const Filter = styled.div`
-  position: absolute;
-  backdrop-filter: blur(32px);
-  background: ${cVar('colorBackgroundOverlay')};
-  width: 280px;
-  height: 100%;
-`
-
-export const MemberInfoContainer = styled.div`
-  position: relative;
-  ${paddingStyles}
-
-  border-bottom: 1px solid ${cVar('colorBorderMutedAlpha')};
-`
-
-export const MemberHandleText = styled(Text)`
-  word-break: break-word;
-`
-
-export const ChannelsSectionTitle = styled(Text)`
-  padding: ${sizes(2)} ${sizes(4)};
-  display: block;
-`
-
-export const SwitchMemberItemListContainer = styled.div`
-  padding: ${sizes(2)} 0;
-`
-
-export const UserBalance = styled.div`
-  display: grid;
-  grid-auto-flow: column;
-  grid-auto-columns: max-content;
-  align-items: center;
-  gap: 5px;
-  margin-top: ${sizes(2)};
-`
-
-export const BalanceContainer = styled.div`
-  clear: both;
-  margin-top: ${sizes(1)};
-  display: grid;
-  grid-auto-flow: column;
-  grid-auto-columns: max-content;
-  align-items: center;
-  gap: 8px;
-`
-
-export const Divider = styled.div`
-  width: 1px;
-  height: 16px;
-  background: ${cVar('colorBorderMutedAlpha')};
-`
-
-export const TextLink = styled(Text)`
-  text-decoration: none;
-  cursor: pointer;
-  display: flex;
-
-  &:hover {
-    text-decoration: underline;
-  }
-
-  svg {
-    margin-left: ${sizes(1)};
-
-    path {
-      fill: ${cVar('colorCoreNeutral200Lighten')};
-    }
-  }
-`
-
-export const AnimatedTextLink = styled(animated(TextLink))``
