+++ packages/atlas/src/components/_overlays/MemberDropdown/MemberDropdown.tsx
@@ -1,11 +1,22 @@
-import { forwardRef, useEffect, useRef, useState } from 'react'
+import bezier from 'bezier-easing'
+import { BN } from 'bn.js'
+import { forwardRef, useCallback, useEffect, useRef, useState } from 'react'
 import mergeRefs from 'react-merge-refs'
+import { useNavigate } from 'react-router'
 import { useTransition } from 'react-spring'
 import useResizeObserver from 'use-resize-observer'
+
+import { absoluteRoutes } from '@/config/routes'
+import { useMemberAvatar } from '@/providers/assets'
+import { useSubscribeAccountBalance } from '@/providers/joystream'
+import { useUser } from '@/providers/user'
+import { cVar } from '@/styles'
 
 import { Container, InnerContainer, SlideAnimationContainer } from './MemberDropdown.styles'
 import { MemberDropdownList } from './MemberDropdownList'
 import { MemberDropdownNav } from './MemberDropdownNav'
+
+import { SendFundsDialog, WithdrawFundsDialog } from '../SendTransferDialogs'
 
 export type MemberDropdownProps = {
   isActive: boolean
@@ -16,18 +27,57 @@ export type MemberDropdownProps = {
 
 export const MemberDropdown = forwardRef<HTMLDivElement, MemberDropdownProps>(
   ({ publisher, isActive, closeDropdown, onChannelChange }, ref) => {
+    const navigate = useNavigate()
+    const { channelId, activeMembership, memberships, signOut, setActiveUser, setSignInModalOpen } = useUser()
+    const [showWithdrawDialog, setShowWithdrawDialog] = useState(false)
+    const [disableScrollDuringAnimation, setDisableScrollDuringAnimation] = useState(false)
+
+    const [showSendDialog, setShowSendDialog] = useState(false)
+    const { url: memberAvatarUrl } = useMemberAvatar(activeMembership)
+    const selectedChannel = activeMembership?.channels.find((chanel) => chanel.id === channelId)
+
+    const { accountBalance, lockedAccountBalance } = useSubscribeAccountBalance()
+    const { accountBalance: channelBalance } = useSubscribeAccountBalance(selectedChannel?.rewardAccount) || new BN(0)
+
+    const hasOneMember = memberships?.length === 1
+
     const [dropdownType, setDropdownType] = useState<'channel' | 'member'>('member')
     const [isList, setIsList] = useState(false)
-    const { ref: measureContainerRef, height: containerHeight = 0 } = useResizeObserver<HTMLDivElement>({
+    const { ref: measureContainerRef, height: measureContainerHeight } = useResizeObserver<HTMLDivElement>({
       box: 'border-box',
     })
     const containerRef = useRef<HTMLDivElement>(null)
 
     const transition = useTransition(isList, {
-      from: { opacity: 0, x: dropdownType === 'channel' ? 280 : -280, position: 'absolute' as const },
+      from: { opacity: 0, x: isList ? 280 : -280, position: 'absolute' as const },
       enter: { opacity: 1, x: 0 },
-      leave: { opacity: 0, x: dropdownType === 'channel' ? -280 : 280 },
+      leave: { opacity: 0, x: isList ? -280 : 280 },
+      config: {
+        duration: parseInt(cVar('animationTimingMedium', true)),
+        // 'animationEasingMedium'
+        easing: bezier(0.03, 0.5, 0.25, 1),
+      },
+      onStart: () => setDisableScrollDuringAnimation(true),
+      onDestroyed: () => setDisableScrollDuringAnimation(false),
     })
+
+    const handleAddNewMember = useCallback(() => {
+      closeDropdown?.()
+      setSignInModalOpen(true)
+    }, [closeDropdown, setSignInModalOpen])
+
+    const handleMemberChange = useCallback(
+      (memberId: string, accountId: string, channelId: string | null) => {
+        setActiveUser({ accountId, memberId, channelId })
+        setIsList(false)
+        closeDropdown?.()
+
+        if (publisher) {
+          navigate(absoluteRoutes.studio.index())
+        }
+      },
+      [closeDropdown, navigate, publisher, setActiveUser]
+    )
 
     useEffect(() => {
       if (!isActive) {
@@ -47,22 +97,44 @@ export const MemberDropdown = forwardRef<HTMLDivElement, MemberDropdownProps>(
         document.removeEventListener('click', handleClickOutside, true)
       }
     }, [closeDropdown, isActive])
+    const toggleWithdrawDialog = () => setShowWithdrawDialog((prevState) => !prevState)
+    const toggleSendDialog = () => setShowSendDialog((prevState) => !prevState)
 
     return (
       <>
+        <WithdrawFundsDialog
+          avatarUrl={memberAvatarUrl}
+          activeMembership={activeMembership}
+          show={showWithdrawDialog}
+          onExitClick={toggleWithdrawDialog}
+          accountBalance={accountBalance}
+          channelBalance={channelBalance}
+          channelId={channelId}
+        />
+        <SendFundsDialog show={showSendDialog} onExitClick={toggleSendDialog} accountBalance={accountBalance} />
         <Container ref={mergeRefs([ref, containerRef])}>
-          <InnerContainer isActive={isActive} containerHeight={containerHeight}>
+          <InnerContainer isActive={isActive} containerHeight={measureContainerHeight}>
             {transition((style, isList) => (
-              <SlideAnimationContainer style={style}>
+              <SlideAnimationContainer style={style} disableVerticalScroll={disableScrollDuringAnimation}>
                 {!isList ? (
                   <div ref={measureContainerRef}>
                     <MemberDropdownNav
-                      switchDropdownType={setDropdownType}
-                      switchToList={(type) => {
+                      channelId={channelId}
+                      onSignOut={signOut}
+                      onShowFundsDialog={() =>
+                        dropdownType === 'channel' ? setShowWithdrawDialog(true) : setShowWithdrawDialog(true)
+                      }
+                      accountBalance={accountBalance}
+                      channelBalance={channelBalance}
+                      lockedAccountBalance={lockedAccountBalance}
+                      activeMembership={activeMembership}
+                      hasOneMember={hasOneMember}
+                      onSwitchDropdownType={setDropdownType}
+                      onSwitchToList={(type) => {
                         setIsList(true)
                         setDropdownType(type)
                       }}
-                      closeDropdown={closeDropdown}
+                      onCloseDropdown={closeDropdown}
                       publisher={publisher}
                       type={dropdownType}
                     />
@@ -70,10 +142,14 @@ export const MemberDropdown = forwardRef<HTMLDivElement, MemberDropdownProps>(
                 ) : (
                   <div ref={measureContainerRef}>
                     <MemberDropdownList
-                      publisher={publisher}
+                      channelId={channelId}
+                      memberships={memberships}
+                      activeMembership={activeMembership}
+                      onMemberChange={handleMemberChange}
                       onChannelChange={onChannelChange}
-                      closeDropdown={closeDropdown}
-                      switchToNav={(type) => {
+                      onCloseDropdown={closeDropdown}
+                      onAddNewMember={handleAddNewMember}
+                      onSwitchToNav={(type) => {
                         setDropdownType(type)
                         setIsList(false)
                       }}
