+++ packages/atlas/src/components/_overlays/MemberDropdown/MemberDropdownList.styles.ts
@@ -1,0 +1,7 @@
+import styled from '@emotion/styled'
+
+import { sizes } from '@/styles'
+
+export const SwitchMemberItemListContainer = styled.div`
+  padding: ${sizes(2)} 0;
+`
